adjustment for mingw64
Try/OpenGothic#27

+++ Engine/gapi/vulkan/vcommandbuffer.cpp
@@ -278,13 +278,13 @@ void VCommandBuffer::copy(VTexture &dest, size_t width, size_t height, size_t mi
   region.bufferRowLength   = 0;
   region.bufferImageHeight = 0;
   region.imageSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-  region.imageSubresource.mipLevel = mip;
+  region.imageSubresource.mipLevel = uint32_t(mip);
   region.imageSubresource.baseArrayLayer = 0;
   region.imageSubresource.layerCount = 1;
   region.imageOffset = {0, 0, 0};
   region.imageExtent = {
-      width,
-      height,
+      uint32_t(width),
+      uint32_t(height),
       1
   };
